Set up prop-types validation for model objects.
