+++ config/mocha-config.js
@@ -4,7 +4,7 @@
 
 const JSDOM = require('jsdom').JSDOM;
 
-process.env.NODE_ENV = 'test';
+process.env.NODE_ENV = 'development';
 
 require('babel-register')({
   "presets": [
@@ -19,7 +19,9 @@ require('babel-register')({
     "istanbul",
     "transform-decorators-legacy",
     "transform-class-properties",
-    "transform-runtime",
+    [
+      "transform-runtime", { "polyfill": false } // Fix for https://github.com/babel/babel/issues/2877
+    ],
     "transform-object-rest-spread"
   ]
 })
