+++ config/webpack.config.showroom.js
@@ -38,7 +38,9 @@ module.exports = {
           plugins: [
             "transform-decorators-legacy",
             "transform-class-properties",
-            "transform-runtime",
+            [
+              "transform-runtime", { "polyfill": false } // Fix for https://github.com/babel/babel/issues/2877
+            ],
             "transform-object-rest-spread"
           ]
         },
