+++ src/crudeditor-lib/lib/index.js
@@ -1,0 +1,1 @@
+export * from './lib';
