+++ src/crudeditor-lib/lib/object-validation/propTypes/formLayout.js
@@ -1,0 +1,33 @@
+import PropTypes from 'prop-types';
+import { uiTypes } from '../lib';
+
+const fieldPropTypes = PropTypes.shape({
+  field: PropTypes.string.isRequired,
+  readOnly: PropTypes.bool.isRequired,
+  render: PropTypes.shape({
+    component: PropTypes.func.isRequired,
+    props: PropTypes.object,
+    value: PropTypes.shape({
+      converter: PropTypes.shape({
+        format: PropTypes.func,
+        parse: PropTypes.func
+      }),
+      propName: PropTypes.string,
+      type: PropTypes.oneOf(uiTypes)
+    })
+  })
+})
+
+export default {
+  formLayout: PropTypes.arrayOf(PropTypes.arrayOf(
+    PropTypes.oneOfType([
+      fieldPropTypes,
+      PropTypes.arrayOf(
+        PropTypes.oneOfType([
+          fieldPropTypes,
+          PropTypes.array
+        ])
+      )
+    ])
+  ))
+}
